+++ lib/grax/schema/custom_field.ex
@@ -3,11 +3,13 @@ defmodule Grax.Schema.CustomField do
 
   defstruct [:name, :from_rdf, :default]
 
-  def new(name, opts) when is_atom(name) do
+  alias Grax.Schema.DataProperty
+
+  def new(schema, name, opts) when is_atom(name) do
     struct!(__MODULE__,
       name: name,
       default: opts[:default],
-      from_rdf: opts[:from_rdf]
+      from_rdf: DataProperty.normalize_custom_mapping_fun(opts[:from_rdf], schema)
     )
   end
 end